requirements.txt
Django==1.9
Pygments==2.1.3
argparse==1.2.1
coreapi==1.32.0
djangorestframework==3.4.3
gunicorn==19.6.0
httpie==0.9.6
itypes==1.1.0
requests==2.11.0
simplejson==3.8.2
uritemplate==0.6
wsgiref==0.1.2
